requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1
groq==0.18.0
anthropic>=0.25.0
